# apps/api/internal/dto/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package openapi

// AdminVerifyRequest defines model for AdminVerifyRequest.
type AdminVerifyRequest struct {
	Token string `json:"token"`
}

// CreateJourneyRequest defines model for CreateJourneyRequest.
type CreateJourneyRequest struct {
	Content   string  `json:"content"`
	Highlight bool    `json:"highlight"`
	Location  string  `json:"location"`
	Name      string  `json:"name"`
	Thumbnail *string `json:"thumbnail,omitempty"`
	Timestamp string  `json:"timestamp"`
}

// CreateJourneyResponse defines model for CreateJourneyResponse.
type CreateJourneyResponse struct {
	Id int64 `json:"id"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// Journey defines model for Journey.
type Journey struct {
	Id        int64   `json:"id"`
	Location  string  `json:"location"`
	Name      string  `json:"name"`
	Thumbnail *string `json:"thumbnail"`
	Timestamp string  `json:"timestamp"`
}

// JourneyContent defines model for JourneyContent.
type JourneyContent struct {
	Content   string `json:"content"`
	CreatedAt string `json:"created_at"`
	Name      string `json:"name"`
	Timestamp string `json:"timestamp"`
}

// JourneysResponse defines model for JourneysResponse.
type JourneysResponse = []Journey

// UploadJourneyImagesResponse defines model for UploadJourneyImagesResponse.
type UploadJourneyImagesResponse = []UploadedImage

// UploadedImage defines model for UploadedImage.
type UploadedImage struct {
	Url string `json:"url"`
}

// UploadJourneyImagesMultipartBody defines parameters for UploadJourneyImages.
type UploadJourneyImagesMultipartBody struct {
	Images []string `json:"images"`
}

// VerifyAdminJSONRequestBody defines body for VerifyAdmin for application/json ContentType.
type VerifyAdminJSONRequestBody = AdminVerifyRequest

// CreateJourneyJSONRequestBody defines body for CreateJourney for application/json ContentType.
type CreateJourneyJSONRequestBody = CreateJourneyRequest

// UploadJourneyImagesMultipartRequestBody defines body for UploadJourneyImages for multipart/form-data ContentType.
type UploadJourneyImagesMultipartRequestBody UploadJourneyImagesMultipartBody