# shared/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: tSR-pg API
  version: 0.1.0
paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check endpoint
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /journeys:
    get:
      operationId: getJourneys
      summary: List journeys
      responses:
        '200':
          description: Journeys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneysResponse'
  /journey:
    post:
      operationId: createJourney
      summary: Create journey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJourneyRequest'
      responses:
        '201':
          description: Created journey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJourneyResponse'
        '400':
          description: Invalid request
        '403':
          description: Invalid admin token
  /journey/images:
    post:
      operationId: uploadJourneyImages
      summary: Upload journey images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - images
              properties:
                images:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Uploaded journey images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadJourneyImagesResponse'
        '400':
          description: Invalid request
        '403':
          description: Invalid admin token
  /journey/{id}:
    get:
      operationId: getJourney
      summary: Get journey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Journey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyContent'
        '404':
          description: Journey not found
  /admin-verify:
    get:
      operationId: getAdminVerify
      summary: Verify admin session
      responses:
        '200':
          description: Admin session accepted
        '403':
          description: Invalid admin session
    post:
      operationId: verifyAdmin
      summary: Verify admin token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminVerifyRequest'
      responses:
        '200':
          description: Admin token accepted
        '403':
          description: Invalid admin token
components:
  schemas:
    AdminVerifyRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    CreateJourneyRequest:
      type: object
      required:
        - name
        - timestamp
        - location
        - content
        - highlight
      properties:
        name:
          type: string
        timestamp:
          type: string
        location:
          type: string
        thumbnail:
          type: string
          nullable: true
        content:
          type: string
        highlight:
          type: boolean
    CreateJourneyResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
    UploadedImage:
      type: object
      required:
        - url
      properties:
        url:
          type: string
    UploadJourneyImagesResponse:
      type: array
      items:
        $ref: '#/components/schemas/UploadedImage'
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    Journey:
      type: object
      required:
        - id
        - name
        - timestamp
        - location
        - thumbnail
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        timestamp:
          type: string
        location:
          type: string
        thumbnail:
          type: string
          nullable: true
    JourneysResponse:
      type: array
      items:
        $ref: '#/components/schemas/Journey'
    JourneyContent:
      type: object
      required:
        - name
        - timestamp
        - created_at
        - content
      properties:
        name:
          type: string
        timestamp:
          type: string
        created_at:
          type: string
        content:
          type: string